fix: change start_variable_number from u8 to u16 for extended variable support
- Change start_variable_number parameter type from u8 to u16 in all plural variable commands
- Update protocol layer: ReadMultipleByteVariables, WriteMultipleByteVariables, ReadMultipleIntegerVariables, WriteMultipleIntegerVariables
- Update client API: read_multiple_byte_variables, write_multiple_byte_variables, read_multiple_integer_variables, write_multiple_integer_variables
- Update mock server: MockState variables HashMap, get_variable/set_variable methods, and all handlers
- Update config: MockServerConfig variables HashMap type
- Remove unnecessary u8::try_from conversions in handlers since instance is already u16
- Support extended variable settings where variable numbers can exceed 99
- Maintain backward compatibility for standard settings (0-99)
- All tests pass with 226 tests successful

+++ moto-hses-client/src/protocol.rs
@@ -78,7 +78,7 @@ impl HsesClient {
     /// Returns an error if communication fails or parameters are invalid
     pub async fn read_multiple_byte_variables(
         &self,
-        start_variable_number: u8,
+        start_variable_number: u16,
         count: u32,
     ) -> Result<Vec<u8>, ClientError> {
         let command = ReadMultipleByteVariables::new(start_variable_number, count)?;
@@ -127,7 +127,7 @@ impl HsesClient {
     /// Returns an error if communication fails or parameters are invalid
     pub async fn write_multiple_byte_variables(
         &self,
-        start_variable_number: u8,
+        start_variable_number: u16,
         values: Vec<u8>,
     ) -> Result<(), ClientError> {
         let command = WriteMultipleByteVariables::new(start_variable_number, values)?;
@@ -151,7 +151,7 @@ impl HsesClient {
     /// Returns an error if communication fails or parameters are invalid
     pub async fn read_multiple_integer_variables(
         &self,
-        start_variable_number: u8,
+        start_variable_number: u16,
         count: u32,
     ) -> Result<Vec<i16>, ClientError> {
         let command = ReadMultipleIntegerVariables::new(start_variable_number, count)?;
@@ -204,7 +204,7 @@ impl HsesClient {
     /// Returns an error if communication fails or parameters are invalid
     pub async fn write_multiple_integer_variables(
         &self,
-        start_variable_number: u8,
+        start_variable_number: u16,
         values: Vec<i16>,
     ) -> Result<(), ClientError> {
         let command = WriteMultipleIntegerVariables::new(start_variable_number, values)?;
